home-manager: add `init` command to main tool
The init command is essentially the old install script but integrated
into the home-manager tool. This simplifies things slightly since we
can use the existing code infrastructure.

The init command is Nix flake aware in the sense that, if we detect
that the user's Nix setup supports Flakes, then we also create an
initial `flake.nix` file.

Finally, we update the installation instructions for the Nix Flakes
standalone setup to use the new init command.